TCC/requirements.txt
langchain
langchain_core
langchain-community
langchain_openai
chromadb
pandas
python-dotenv
selenium
unstructured
ragas
termcolor
pydantic
streamlit
streamlit-float
streamlit-chat
langchain-chroma
pdfplumber
gdown
s3fs
st-files-connection
langgraph